[김지협] sprint3

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicMessageService.java
@@ -1,51 +1,57 @@
 package com.sprint.mission.discodeit.service.basic;
 
+import com.sprint.mission.discodeit.dto.CreateMessageDto;
+import com.sprint.mission.discodeit.dto.UpdateMessageDto;
+import com.sprint.mission.discodeit.entity.BinaryContent;
 import com.sprint.mission.discodeit.entity.Message;
+import com.sprint.mission.discodeit.repository.BinaryContentRepository;
+import com.sprint.mission.discodeit.repository.ChannelRepository;
 import com.sprint.mission.discodeit.repository.MessageRepository;
+import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.service.ChannelService;
 import com.sprint.mission.discodeit.service.MessageService;
 import com.sprint.mission.discodeit.service.UserService;
+import lombok.NoArgsConstructor;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.UUID;
 import java.util.concurrent.atomic.AtomicInteger;
 
+@Service
+@RequiredArgsConstructor
 public class BasicMessageService implements MessageService {
     private final MessageRepository messageRepository;
-    private final UserService userService;
-    private final ChannelService channelService;
-    private final AtomicInteger messageNum = new AtomicInteger(1);
-
-    public BasicMessageService(MessageRepository messageRepository, UserService userService, ChannelService channelService) {
-        this.messageRepository = messageRepository;
-        this.userService = userService;
-        this.channelService = channelService;
-    }
+    private final UserRepository userRepository;
+    private final ChannelRepository channelRepository;
+    private final BinaryContentRepository binaryContentRepository;
 
     @Override
-    public Message create(String content, UUID userKey, UUID channelKey) {
-        if (content == null || content.isEmpty()) {
-            throw new IllegalArgumentException("[Error] 내용을 입력해주세요");
+    public Message create(CreateMessageDto messageDto) {
+        if (!channelRepository.existsByKey(messageDto.channelKey())) {
+            throw new NoSuchElementException("[Error] 채널이 존재하지 않습니다. " + messageDto.channelKey());
         }
-
-        String username = userService.getUserName(userKey);
-        String channelName = channelService.getChannelName(channelKey);
-        int messageId = messageNum.getAndIncrement();
-
-        Message message = new Message(messageId, content, userKey, channelKey, username, channelName);
+        if (!userRepository.existsByKey(messageDto.userKey())) {
+            throw new NoSuchElementException("[Error] 유저가 존재하지 않습니다. " + messageDto.userKey());
+        }
+        Message message = new Message(messageDto.content(), messageDto.userKey(), messageDto.channelKey(), messageDto.attachmentKeys());
         messageRepository.save(message);
         return message;
     }
 
     @Override
-    public Message read(UUID channelKey) {
-        return messageRepository.findAllByChannelKey(channelKey).stream()
-                .max((m1, m2) -> Long.compare(m1.getCreatedAt(), m2.getCreatedAt()))
-                .orElseThrow(() -> new IllegalStateException("[Error] 읽을 메시지가 없습니다."));
+    public Message read(UUID messageKey) {
+        Message message = messageRepository.findByKey(messageKey);
+        if (message == null) {
+            throw new IllegalStateException("[Error] 읽을 메시지가 없습니다.");
+        }
+        return message;
     }
 
     @Override
-    public List<Message> readAll(UUID channelKey) {
+    public List<Message> readAllByChannelKey(UUID channelKey) {
         List<Message> messages = messageRepository.findAllByChannelKey(channelKey);
         if (messages.isEmpty()) {
             throw new IllegalStateException("[Error] 읽을 메시지가 없습니다.");
@@ -54,22 +60,28 @@ public class BasicMessageService implements MessageService {
     }
 
     @Override
-    public UUID update(int messageId, String content) {
-        UUID messageKey = messageRepository.findKeyByMessageId(messageId);
-        if (messageKey == null) {
+    public UpdateMessageDto update(UpdateMessageDto dto) {
+        Message message = messageRepository.findByKey(dto.messageKey());
+        if (message == null) {
             throw new IllegalArgumentException("[Error] 해당 메시지가 존재하지 않습니다");
         }
-
-        Message message = messageRepository.findByKey(messageKey);
-        if (!content.isEmpty()) {
-            message.updateContent(content);
+        if (!dto.content().isEmpty()) {
+            message.updateContent(dto.content());
         }
-        return messageKey;
+        return new UpdateMessageDto(message.getUuid(), message.getContent());
     }
 
     @Override
-    public void delete(int messageId) {
-        UUID messageKey = messageRepository.findKeyByMessageId(messageId);
+    public void delete(UUID messageKey) {
+        Message message = messageRepository.findByKey(messageKey);
+        if (message == null) {
+            throw new IllegalArgumentException("[Error] 해당 메시지가 존재하지 않습니다");
+        }
+        List<UUID> attachmentKeys = message.getAttachmentKeys();
+        for (UUID attachmentKey : attachmentKeys) {
+            binaryContentRepository.delete(attachmentKey);
+        }
+
         messageRepository.delete(messageKey);
     }
 }
